+++ source/scene/compartments/sector.cpp
@@ -5,7 +5,7 @@
 
 #include "sector.h"
 
-Sector::Sector(Scene* scene, float x, float z)
+Sector::Sector(Scene* scene, int x, int z)
     :scene(scene), x(x), z(z), biomes_processed(false), mesh_rebuilt(false)
 {
     for (unsigned int _x = 0; _x < SECTOR_HORIZONTAL_SIZE; _x++)
@@ -46,7 +46,7 @@ Segment* Sector::CreateSegment(const SegmentIndex& index)
     if (!_result)
     {
         auto _segment = new Segment(scene, BlockType::TEST, false,
-            x + index.x * SEGMENT_LENGTH, (float)(index.y * SEGMENT_LENGTH), z + index.z * SEGMENT_LENGTH);
+            x + index.x * SEGMENT_LENGTH, index.y * SEGMENT_LENGTH, z + index.z * SEGMENT_LENGTH);
         segments[index.x][index.y][index.z].store(_segment);
         mesh_rebuilt.store(false);
         return _segment;
@@ -67,7 +67,7 @@ Segment* Sector::GetSegment(const SegmentIndex& index, bool force)
     else if (force)
     {
         auto _segment = new Segment(scene, BlockType::TEST, false,
-            x + index.x * SEGMENT_LENGTH, (float)(index.y * SEGMENT_LENGTH), z + index.z * SEGMENT_LENGTH);
+            x + index.x * SEGMENT_LENGTH, index.y * SEGMENT_LENGTH, z + index.z * SEGMENT_LENGTH);
         segments[index.x][index.y][index.z].store(_segment);
         mesh_rebuilt.store(false);
         return _segment;
